add login and refactor routes
