+++ src/lib.rs
@@ -17,14 +17,15 @@ pub fn run(
     let server = HttpServer::new(move || {
         App::new()
             .wrap(TracingLogger::default())
-            .route(
-                "/health_check",
-                web::get().to(routes::health_check::health_check),
-            )
-            .route("/home", web::get().to(routes::home_page::home_page))
-            .route("/sign_up", web::post().to(routes::sign_up::sign_up))
-            .route("/login", web::post().to(routes::login::sign_in))
-            .route("/me", web::get().to(routes::login::get_me_handler))
+            .configure(routes::handler::config)
+            // .route(
+            //     "/health_check",
+            //     web::get().to(routes::health_check::health_check),
+            // )
+            // .route("/home", web::get().to(routes::home_page::home_page))
+            // .route("/sign_up", web::post().to(routes::sign_up::sign_up))
+            // .route("/login", web::post().to(routes::login::sign_in))
+            // .route("/me", web::get().to(routes::login::get_me_handler))
             .app_data(web::Data::new(config::AppState {
                 db: connection.clone(),
                 config: config.clone(),
