+++ src/routes/health_check.rs
@@ -1,5 +1,6 @@
-use actix_web::{HttpResponse, Responder};
+use actix_web::{HttpResponse, get, Responder};
 
+#[get("/health_check")]
 pub async fn health_check() -> impl Responder {
     tracing::info!("testing health check");
     HttpResponse::Ok().finish()
