+++ src/routes/mod.rs
@@ -1,4 +1,4 @@
 pub mod health_check;
 pub mod home_page;
-pub mod login;
-pub mod sign_up;
+pub mod auth;
+pub mod handler;
